Adicion de Auditoria de procesos y ajustes tarifa operaciones y vuelos charter

+++ src/main/java/com/ath/adminefectivo/dto/TarifasOperacionDTO.java
@@ -47,7 +47,7 @@ public class TarifasOperacionDTO {
 
 	private String comisionAplicar;
 
-	private int valorTarifa; 
+	private Double valorTarifa; 
 
 	private int estado;
